Refactored create meeting form and added additional fields
